-refactor [back] entities extends abstract class BaseEntity; -fix: [back] Month.month has validators;
