+++ backend/package.json
@@ -26,6 +26,8 @@
     "@nestjs/platform-express": "^9.0.0",
     "@nestjs/swagger": "^6.1.3",
     "@nestjs/typeorm": "^9.0.1",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.13.2",
     "pg": "^8.8.0",
     "psql": "^0.0.1",
     "reflect-metadata": "^0.1.13",
