+++ backend/src/DTOs/year.ts
@@ -1,4 +1,4 @@
-import BaseDTO from "../database/bases";
+import BaseDTO from "src/database/BaseDTO";
 import MonthDTO from "./month";
 
 export default interface YearDTO extends BaseDTO {
