+++ backend/src/database/BaseDTO.ts
@@ -1,0 +1,5 @@
+export default interface BaseDTO {
+  id: string
+  createdAt: Date
+  updatedAt: Date
+}
