+++ backend/src/entities/Year.ts
@@ -1,21 +1,16 @@
 import { Service } from "typedi";
-import { v4 as uuid } from "uuid";
 import {
   Entity,
   Column,
-  PrimaryColumn,
-  CreateDateColumn,
-  UpdateDateColumn,
-  DeleteDateColumn,
   OneToMany,
 } from "typeorm";
-import { OneToManyOptions } from "src/database/bases";
+import { BaseEntity, OneToManyOptions } from "src/database/BaseEntity";
 import { Month } from "./Month";
 import YearDTO from "src/DTOs/year";
 
 @Service()
 @Entity("years")
-export class Year  {
+export class Year extends BaseEntity {
   // COLUMNS
   @Column()
   year: string
@@ -23,23 +18,6 @@ export class Year  {
   // RELATIONS
   @OneToMany(() => Month, month => month.year, OneToManyOptions)
   months: Month[]
-
-  // BASE
-  @PrimaryColumn({ type: 'uuid' })
-  id: string;
-
-  @CreateDateColumn()
-  createdAt: Date
-
-  @UpdateDateColumn()
-  updatedAt: Date
-
-  @DeleteDateColumn()
-  deletedAt: Date
-
-  constructor() {
-    if(!this.id) this.id = uuid()
-  }
 
   public static toDTO(row: Year): YearDTO {
     return {
